+++ packages/app-desktop/gui/NoteListItem/utils/types.ts
@@ -1,0 +1,3 @@
+import * as React from 'react';
+
+export type OnCheckboxChange = (event: React.ChangeEvent<HTMLInputElement>)=> void;
